feat: Add .eslintrc.cjs configuration file
This commit adds a new configuration file, .eslintrc.cjs, which includes environment settings for browser and ES2021. It extends the "standard-with-typescript" ruleset and overrides it for files with the extension .eslintrc.js or .eslintrc.cjs. The parser options are set to use the latest ECMAScript version and source type as module. No specific rules are defined in this commit.

BREAKING CHANGE: Removed .eslintrc.js configuration file